pyproject.toml
[tool.poetry]
name = "wordwaybot"
version = "0.1.0"
description = ""
authors = ["Олег Латыш <olatysh@migcredit.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.7"
python-telegram-bot = "^21.4"
dynaconf = "^3.2.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
